+++ Model/Position.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Position: Identifiable {
+    let id: Int
+    var note: String?
+}